Refactor chart generators into dedicated charts subdirectory
- Move all chart-related files into src/reporting/charts/
- Rename chart_generator.py to base.py
- Simplify file names by removing redundant _chart_generator suffix
- Update import paths in all affected files
- Add __init__.py to make charts a proper Python package
- Maintain same class names and functionality
