+++ src/reporting/charts/__init__.py
@@ -1,0 +1,1 @@
+"""Chart generators for health data visualization."""
