+++ src/reporting/charts/macro_ratio.py
@@ -1,5 +1,5 @@
-from .chart_generator import ChartGenerator
-from .reporting_config import ReportingConfig
+from .base import ChartGenerator
+from ..reporting_config import ReportingConfig
 
 
 class MacroRatioChartGenerator(ChartGenerator):
